requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cssselect==1.1.0
Deprecated==1.2.13
fake-useragent==0.1.11
Flask==2.1.2
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
lxml==4.8.0
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
parsel==1.6.0
pyparsing==3.0.9
redis==4.3.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
w3lib==1.22.0
Werkzeug==2.1.2
win32-setctime==1.1.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
